moved query function to where it should be
