+++ packages/rs-drive/src/drive/contract/mod.rs
@@ -155,31 +155,6 @@ pub struct ContractFetchInfo {
 
 #[cfg(feature = "full")]
 impl Drive {
-    pub(crate) fn fetch_contract_history_query(
-        contract_id: [u8; 32],
-        start_at_date: u64,
-        limit: Option<u16>,
-        offset: Option<u16>,
-    ) -> Result<PathQuery, Error> {
-        let limit = limit.unwrap_or_else(|| MAX_CONTRACT_HISTORY_FETCH_LIMIT);
-        if limit > MAX_CONTRACT_HISTORY_FETCH_LIMIT || limit < 1 {
-            return Err(Error::Drive(DriveError::InvalidContractHistoryFetchLimit(
-                limit,
-            )));
-        }
-
-        let query = Query::new_single_query_item_with_direction(
-            QueryItem::RangeAfter(std::ops::RangeFrom {
-                start: encode_u64(start_at_date),
-            }),
-            false,
-        );
-
-        Ok(PathQuery::new(
-            paths::contract_keeping_history_storage_path_vec(&contract_id),
-            SizedQuery::new(query, Some(limit), offset),
-        ))
-    }
     /// Adds a contract to storage.
     fn add_contract_to_storage(
         &self,
