+++ packages/rs-drive/src/drive/contract/queries.rs
@@ -4,7 +4,11 @@ use crate::drive::contract::paths::{
 use crate::drive::Drive;
 use crate::error::Error;
 use crate::error::Error::GroveDB;
-use grovedb::PathQuery;
+use grovedb::{PathQuery, SizedQuery};
+use crate::common::encode::encode_u64;
+use crate::drive::contract::{MAX_CONTRACT_HISTORY_FETCH_LIMIT, paths};
+use crate::error::drive::DriveError;
+use crate::query::{Query, QueryItem};
 
 impl Drive {
     /// Creates a path query for a specified contract.
@@ -66,5 +70,50 @@ impl Drive {
             queries.push(Self::fetch_contract_query(*contract_id));
         }
         PathQuery::merge(queries.iter().collect()).map_err(GroveDB)
+    }
+
+    /// Creates a path query for historical entries of a specified contract.
+    ///
+    /// This function takes a slice of contract IDs and creates a path query for fetching
+    /// the historical data of the specified contract.
+    ///
+    /// # Arguments
+    ///
+    /// * `contract_id` - A contract ID as a 32-byte array. The contract ID is used to
+    ///   create the path query.
+    ///
+    /// # Returns
+    ///
+    /// * `Result<PathQuery, Error>` - If successful, returns a `PathQuery` object representing the
+    ///   merged query for fetching the contracts' data. If limit are outside of the
+    ///   allowed range, returns an `Error`.
+    ///
+    /// # Errors
+    ///
+    /// This function returns an error if the limit is out of the allowed range.
+    pub fn fetch_contract_history_query(
+        contract_id: [u8; 32],
+        start_at_date: u64,
+        limit: Option<u16>,
+        offset: Option<u16>,
+    ) -> Result<PathQuery, Error> {
+        let limit = limit.unwrap_or_else(|| MAX_CONTRACT_HISTORY_FETCH_LIMIT);
+        if limit > MAX_CONTRACT_HISTORY_FETCH_LIMIT || limit < 1 {
+            return Err(Error::Drive(DriveError::InvalidContractHistoryFetchLimit(
+                limit,
+            )));
+        }
+
+        let query = Query::new_single_query_item_with_direction(
+            QueryItem::RangeAfter(std::ops::RangeFrom {
+                start: encode_u64(start_at_date),
+            }),
+            false,
+        );
+
+        Ok(PathQuery::new(
+            paths::contract_keeping_history_storage_path_vec(&contract_id),
+            SizedQuery::new(query, Some(limit), offset),
+        ))
     }
 }
